:art: add type annotation and mypy config
Added type annotation where needed for mypy.
Some errors are ignored, typically in a more complex flow and dynamic
typing. These are left for later.
Errors are ignored for Visualization as that code will likely get merged
with the Dash app later.

+++ gama/search_methods/async_ea.py
@@ -45,7 +45,7 @@ class AsyncEA(BaseSearch):
         )
         self.output = []
 
-    def dynamic_defaults(self, x: pd.DataFrame, y: pd.DataFrame, time_limit: int):
+    def dynamic_defaults(self, x: pd.DataFrame, y: pd.DataFrame, time_limit: float):
         pass
 
     def search(self, operations: OperatorSet, start_candidates: List[Individual]):
